+++ src/hooks/template/useBadge.ts
@@ -1,0 +1,22 @@
+import { useMenuGetters } from '@/store'
+
+export function useBadge() {
+  const { getMenuOptions } = useMenuGetters()
+
+  const hidden = () => {}
+
+  const show = () => {}
+
+  const toggle = () => {}
+
+  const updateLabel = () => {}
+
+  return {
+    hidden,
+    show,
+    toggle,
+    updateLabel,
+  }
+}
+
+export type UseBadgeReturnType = ReturnType<typeof useBadge>