+++ src/views/dashboard/index.tsx
@@ -66,16 +66,16 @@ const Dashboard = defineComponent({
         value: 'Vue3.x',
       },
       {
-        label: 'Vite4.0',
-        value: 'Vite4.0',
+        label: 'Vite5.x',
+        value: 'Vite5.x',
       },
       {
         label: 'Pinia',
         value: 'Pinia',
       },
       {
-        label: 'TSX',
-        value: 'TSX',
+        label: 'TS(X)',
+        value: 'TS(X)',
       },
     ]